Move file utils to SafeFiles class
